Moved source file to src/index.js, created multiple output files in dist
