+++ package.json
@@ -24,13 +24,14 @@
   "author": "Steven Kalt",
   "license": "GPL-3.0",
   "devDependencies": {
+    "@turf/helpers": "^6.0.0-beta.3",
     "babel-plugin-external-helpers": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-env": "^1.6.1",
     "eslint": "^4.0.0",
     "eslint-config-google": "^0.8.0",
     "jsdoc-to-markdown": "^3.0.0",
     "mocha": "^4.0.1",
-    "rollup": "^0.42.0",
+    "rollup": "^0.54.1",
     "rollup-plugin-babel": "^2.7.1",
     "rollup-plugin-commonjs": "^8.0.2",
     "rollup-plugin-eslint": "^3.0.0",
